+++ services/rapidfs/model_cache_rule_job_info.go
@@ -1,9 +1,0 @@
-package rapidfs
-
-type CacheRuleJobInfo struct {
-	CacheRuleJobId *string `json:"cacheRuleJobId,omitempty"`
-	Status         *string `json:"status,omitempty"`
-	Size           *int64  `json:"size,omitempty"`
-	StartTime      *string `json:"startTime,omitempty"`
-	EndTime        *string `json:"endTime,omitempty"`
-}