+++ services/rapidfs/model_describe_cache_rule_jobs_response.go
@@ -4,9 +4,9 @@ import "github.com/baidubce/baiducloud-go-sdk/bce"
 
 type DescribeCacheRuleJobsResponse struct {
 	bce.BaseResponse
-	CacheRuleJobInfos []*CacheRuleJobInfo `json:"cacheRuleJobInfos,omitempty"`
-	Marker            *string             `json:"marker,omitempty"`
-	IsTruncated       *bool               `json:"isTruncated,omitempty"`
-	NextMarker        *string             `json:"nextMarker,omitempty"`
-	MaxKeys           *int32              `json:"maxKeys,omitempty"`
+	CacheJobInfos []*CacheJobInfo `json:"cacheJobInfos,omitempty"`
+	Marker        *string         `json:"marker,omitempty"`
+	IsTruncated   *bool           `json:"isTruncated,omitempty"`
+	NextMarker    *string         `json:"nextMarker,omitempty"`
+	MaxKeys       *int32          `json:"maxKeys,omitempty"`
 }